# routes/swagger-output local.json
{
  "swagger": "2.0",
  "info": {
    "title": "Blog API",
    "description": "A Blog API that allows authors to publish posts and comment on other posts",
    "version": "1.0.0"
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "oauth": {
      "type": "oauth2",
      "authorizationUrl": "http://localhost:3000/users/auth/google",
      "flow": "implicit",
      "scopes": {
        "read:user": "read user data"
      }
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/users/auth/login": {
      "post": {
        "description": "Login a User",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "Login User",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "professor@byui.edu"
                },
                "password": {
                  "type": "string",
                  "example": "Potter11?"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully Logged In"
          },
          "422": {
            "description": "Kindly check the provided data"
          }
        }
      }
    },
...
  }
}